requirements/test.txt
-r base.txt
pytest==5.1.3
pytest-env==0.6.2
pytest-flask==0.15.0
requests-mock==1.8.0